Code written for Caesar Cipher project.
Basic function added that takes a string and a number, and will shift each letter in the string over the specified number of letters further into the alphabet, or rotate back to the beginning if the letter is at the end of the alphabet. Only shifts in one direction with current version.